+++ TeslaCam.sln
@@ -8,8 +8,6 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeslaCam.Tests", "TeslaCam.
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeslaCam.Data", "TeslaCam.Data\TeslaCam.Data.csproj", "{CC8D9C71-FACD-4F9F-89FD-B67C0C58A144}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TeslaCam.Processor", "TeslaCam.Processor\TeslaCam.Processor.csproj", "{14800C6D-1154-45D3-B0F1-883A62279CBE}"
-EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -28,10 +26,6 @@ Global
 		{CC8D9C71-FACD-4F9F-89FD-B67C0C58A144}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{CC8D9C71-FACD-4F9F-89FD-B67C0C58A144}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{CC8D9C71-FACD-4F9F-89FD-B67C0C58A144}.Release|Any CPU.Build.0 = Release|Any CPU
-		{14800C6D-1154-45D3-B0F1-883A62279CBE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{14800C6D-1154-45D3-B0F1-883A62279CBE}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{14800C6D-1154-45D3-B0F1-883A62279CBE}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{14800C6D-1154-45D3-B0F1-883A62279CBE}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
