+++ TeslaCam/PackageManager.cs
@@ -1,12 +1,13 @@
 ﻿using CliWrap;
-using CliWrap.Buffered;
+using Serilog;
 
-namespace TeslaCam.Processor;
+namespace TeslaCam;
 
 public static class PackageManager
 {
     public static async Task<bool> InstallWinGetPackage(string name)
     {
+        Log.Information($"Installing {name} using winget...");
         var wingetInstalled = (await Cli.Wrap("winget")
             .WithArguments("--version")
             .ExecuteAsync())
@@ -14,12 +15,15 @@ public static class PackageManager
 
         if (!wingetInstalled)
         {
+            Log.Error($"winget is not installed");
             return false;
         }
 
         var installResult = await Cli.Wrap("winget")
             .WithArguments(["install", name])
             .ExecuteAsync();
+
+        Log.Information($"Installed: {installResult}");
 
         return installResult.IsSuccess;
     }
@@ -32,12 +36,5 @@ public static class PackageManager
             .IsSuccess;
     }
 
-    public static async Task<string[]> FindFFmpegPaths()
-    {
-        var result = await Cli.Wrap("where")
-            .WithArguments("ffmpeg")
-            .ExecuteBufferedAsync();
-
-        return result.StandardOutput.Split(['\r', '\n'], StringSplitOptions.RemoveEmptyEntries);
-    }
+    public static Task<bool> InstallFFmpeg() => InstallWinGetPackage("Gyan.FFmpeg.Shared");
 }
