+++ TeslaCam/TeslaCam.csproj
@@ -7,6 +7,7 @@
   </PropertyGroup>
 
   <ItemGroup>
+    <PackageReference Include="CliWrap" Version="3.7.0" />
     <PackageReference Include="CommunityToolkit.Mvvm" Version="8.4.0" />
     <PackageReference Include="FFME.Windows" Version="4.4.350" />
     <PackageReference Include="Serilog" Version="4.2.0" />
@@ -16,7 +17,6 @@
 
   <ItemGroup>
     <ProjectReference Include="..\TeslaCam.Data\TeslaCam.Data.csproj" />
-    <ProjectReference Include="..\TeslaCam.Processor\TeslaCam.Processor.csproj" />
   </ItemGroup>
 
 </Project>
